Create home module for building Home tab
This took a bit of research to figure out how to do. I
grabbed svg-inline-loader because I couldn't get my svg file
to load once removing the directly created HTML files and
building it out with Javascript. After some searching, I saw
that you needed the svg-inline-loader to do something like that,
but it still didn't seem to want to appear. Well the reason was this:
When you import the svg file in Javascript, that import is just a path
to the SVG file. So to properly create the SVG for HTML, you build
out your img element, but when you set the attribute for src, it needs
to be what you named the svg's import! After that, it's good!