Remove unnecessary svg-inline-loader dependency
So this whole time, I thought svg-inline-loader was what was loading
my svg, but it turns out it was actually just being loaded as an asset
with html-loader. I tested using svg-inline-loader by removing the |svg|
part in the html-loader module.loaders section, and don't actually want
that kind of importing(loading?).